1142 - Cast partitions to integers

+++ projects/_01_ingest/cqc_api/fargate/cqc_locations_3_full_flattened.py
@@ -88,10 +88,10 @@ def main(
             date_str = str(delta_import_date)
             merged_lf = merged_lf.with_columns(
                 [
-                    pl.lit(date_str).alias(Keys.import_date),
-                    pl.lit(date_str[:4]).alias(Keys.year),
-                    pl.lit(date_str[4:6]).alias(Keys.month),
-                    pl.lit(date_str[6:8]).alias(Keys.day),
+                    pl.lit(date_str).cast(pl.Int64).alias(Keys.import_date),
+                    pl.lit(date_str[:4]).cast(pl.Int64).alias(Keys.year),
+                    pl.lit(date_str[4:6]).cast(pl.Int64).alias(Keys.month),
+                    pl.lit(date_str[6:8]).cast(pl.Int64).alias(Keys.day),
                 ]
             )
